Prepare implementation of Platform
