+++ pkg/podman/podman.go
@@ -2,13 +2,11 @@ package podman
 
 import (
 	"fmt"
-	"io"
 	"os"
 	"os/exec"
 
 	corev1 "k8s.io/api/core/v1"
 	jsonserializer "k8s.io/apimachinery/pkg/runtime/serializer/json"
-	"k8s.io/klog"
 	"k8s.io/kubectl/pkg/scheme"
 )
 
@@ -87,30 +85,5 @@ func (o *PodmanCli) PodRm(podname string) error {
 func (o *PodmanCli) VolumeRm(volumeName string) error {
 	out, err := exec.Command("podman", "volume", "rm", volumeName).Output()
 	fmt.Printf("%s\n", string(out))
-	return err
-}
-
-func (o *PodmanCli) ExecCMDInContainer(containerName, podName string, cmd []string, stdout io.Writer, stderr io.Writer, stdin io.Reader, tty bool) error {
-	options := []string{}
-	if tty {
-		options = append(options, "--tty")
-	}
-
-	name := fmt.Sprintf("%s-%s", podName, containerName)
-
-	args := []string{"exec", "--interactive"}
-	args = append(args, options...)
-	args = append(args, name)
-	args = append(args, cmd...)
-
-	command := exec.Command("podman", args...)
-	command.Stdin = stdin
-
-	klog.V(4).Infof("exec podman %v\n", args)
-	out, err := command.Output()
-	if err != nil {
-		return err
-	}
-	_, err = stdout.Write(out)
 	return err
 }
